fix sofia windows build.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@5704 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/sofia-sip/libsofia-sip-ua/tport/tport_type_tcp.c
@@ -313,7 +313,7 @@ int tport_recv_stream(tport_t *self)
     if (i + self->tp_ping >= 4)
       tport_tcp_pong(self);
     else
-      self->tp_ping += i;
+      self->tp_ping += (unsigned short)i;
 
     if (i == iovec->siv_len && veclen == 1) {
       SU_DEBUG_7(("%s(%p): received %u bytes of keepalive\n",
